+++ partials/front-page/above-the-fold/featured-post-primary.php
@@ -39,9 +39,9 @@
               }
             ?>
           </h5>
-          <p class="fs-3-sans mt-2 mb-0">
+          <div class="fs-3-sans mt-2 mb-0">
             <?php render_short_description($post_id); ?>
-          </p>
+          </div>
         </a>
       </div>
       <?php