+++ partials/front-page/show-blocks/novara-live.php
@@ -53,9 +53,9 @@
             <div class="grid-item is-l-24 is-xxl-8 mt-l-3">
               <a href="<?php the_permalink(); ?>" class="ui-hover">
                 <h6 class="fs-7"><?php the_title(); ?></h6>
-                <p class="fs-4-sans font-weight-regular mt-3">
+                <div class="fs-4-sans font-weight-regular text-paragraph-breaks mt-3">
                   <?php render_short_description($post->ID); ?>
-                </p>
+                </div>
               </a>
               <?php
                 if (!empty($meta['_cmb_related_posts'])) {