+++ partials/front-page/above-the-fold/featured-post-tertiary.php
@@ -27,7 +27,7 @@
           }
         ?>
       </h5>
-      <p class="fs-3-sans mt-1 mb-0">
+      <div class="fs-3-sans mt-1 mb-0">
         <?php
           if ($is_article) {
             render_standfirst($post_id);
@@ -35,6 +35,6 @@
             render_short_description($post_id);
           }
         ?>
-      </p>
+      </div>
     </a>
   </div>